Refactor PasswordModal
