+++ src/renderer/components/modal/confirmation/index.ts
@@ -1,3 +1,4 @@
 export * from './ConfirmationModal'
 export * from './LedgerConfirmationModal'
 export * from './RemoveWalletConfirmationModal'
+export * from './WalletPasswordConfirmationModal'
