+++ src/renderer/components/wallet/txs/send/SendFormBCH.tsx
@@ -28,7 +28,7 @@ import { FeeRD, Memo, SendTxParams } from '../../../../services/chain/types'
 import { AddressValidation, WalletBalances } from '../../../../services/clients'
 import { ValidatePasswordHandler } from '../../../../services/wallet/types'
 import { WalletBalance } from '../../../../services/wallet/types'
-import { PasswordModal } from '../../../modal/password'
+import { WalletPasswordConfirmationModal } from '../../../modal/confirmation'
 import * as StyledR from '../../../shared/form/Radio.styles'
 import { MaxBalanceButton } from '../../../uielements/button/MaxBalanceButton'
 import { UIFeesRD } from '../../../uielements/fees'
@@ -281,7 +281,7 @@ export const SendFormBCH: React.FC<Props> = (props): JSX.Element => {
   const renderPwModal = useMemo(
     () =>
       showPwModal ? (
-        <PasswordModal
+        <WalletPasswordConfirmationModal
           onSuccess={sendHandler}
           onClose={() => setShowPwModal(false)}
           validatePassword$={validatePassword$}
