+++ src/renderer/i18n/de/wallet.ts
@@ -69,7 +69,8 @@ const wallet: WalletMessages = {
     'Der eingegebene Wert sollte nicht höher als Dein Guthaben abzgl. Gebühren sein',
   'wallet.errors.fee.notCovered': 'Die Gebühren sind nicht über Dein Guthaben ({balance}) gedeckt',
   'wallet.errors.invalidChain': 'Invalide Chain: {chain}',
-  'wallet.password.confirmation': 'Password bestätigen',
+  'wallet.password.confirmation.title': 'Wallet Passwort bestätigen',
+  'wallet.password.confirmation.description': 'Bitte gebe zur Bestätigung Dein Wallet Password ein.',
   'wallet.password.confirmation.pending': 'Überprüfe Passwort',
   'wallet.password.confirmation.error': 'Passwort ist falsch',
   'wallet.password.repeat': 'Passwort wiederholen',
