+++ src/renderer/i18n/en/wallet.ts
@@ -66,7 +66,8 @@ const wallet: WalletMessages = {
   'wallet.errors.amount.shouldBeLessThanBalanceAndFee': 'Amount should be less than your balance minus fee',
   'wallet.errors.fee.notCovered': 'Fees are not covered by your balance ({balance})',
   'wallet.errors.invalidChain': 'Invalid chain: {chain}',
-  'wallet.password.confirmation': 'Password confirmation',
+  'wallet.password.confirmation.title': 'Wallet password confirmation',
+  'wallet.password.confirmation.description': 'Please enter your wallet password.',
   'wallet.password.confirmation.pending': 'Validating password',
   'wallet.password.confirmation.error': 'Password is wrong',
   'wallet.password.repeat': 'Repeat password',
