+++ src/renderer/i18n/fr/wallet.ts
@@ -68,7 +68,8 @@ const wallet: WalletMessages = {
   'wallet.errors.amount.shouldBeLessThanBalanceAndFee': 'Le montant doit être inférieur à votre solde moins les frais',
   'wallet.errors.fee.notCovered': 'Les frais ne sont pas couverts par votre solde de ({balance})',
   'wallet.errors.invalidChain': 'Chaîne non valide : {chain}',
-  'wallet.password.confirmation': 'Confirmation du mot de passe',
+  'wallet.password.confirmation.title': 'Confirmation du mot de passe',
+  'wallet.password.confirmation.description': 'Please enter your wallet password. - FR',
   'wallet.password.confirmation.pending': 'Validation du mot de passe',
   'wallet.password.confirmation.error': 'Le mot de passe est erroné',
   'wallet.password.repeat': 'Répétez le mot de passe',
