+++ src/renderer/i18n/ru/wallet.ts
@@ -67,7 +67,8 @@ const wallet: WalletMessages = {
     'Количество должно быть меньше, чем ваш баланс после вычета комисси',
   'wallet.errors.fee.notCovered': ' Комиссия не покрывается вашим банаслом ({balance})',
   'wallet.errors.invalidChain': 'Цепь недействительна',
-  'wallet.password.confirmation': 'Подтверждение пароля',
+  'wallet.password.confirmation.title': 'Подтверждение пароля',
+  'wallet.password.confirmation.description': 'Please enter your wallet password. - RU',
   'wallet.password.confirmation.pending': 'Проверка пароля',
   'wallet.password.confirmation.error': 'Неверный пароль',
   'wallet.password.repeat': 'Повторите парроль',
