+++ src/renderer/i18n/types.ts
@@ -175,7 +175,8 @@ type WalletMessageKey =
   | 'wallet.send.fastest'
   | 'wallet.send.fast'
   | 'wallet.send.average'
-  | 'wallet.password.confirmation'
+  | 'wallet.password.confirmation.title'
+  | 'wallet.password.confirmation.description'
   | 'wallet.password.confirmation.pending'
   | 'wallet.password.confirmation.error'
   | 'wallet.password.repeat'
